+++ projects/igniteui-angular/migrations/update-12_1_0/index.spec.ts
@@ -461,4 +461,195 @@ public onBannerOpened(event: BannerEventArgs) {
 `.replace(lineBreaksAndSpaceRegex, ''));
     });
 
+    // IgxDropDown
+    it('should update IgxDropDown event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update dropdown event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+        <igx-drop-down
+            (onOpening)="handleEvent($event, 'opening')"
+            (onOpened)="handleEvent($event, 'opened')"
+            (onClosing)="handleEvent($event, 'closing')"
+            (onClosed)="handleEvent($event, 'closed')"
+            (onSelection)="handleEvent($event, 'selection')"
+        >
+            Display something onOpening, onClosing, onOpened, onClosed
+        </igx-drop-down>`);
+            const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+                .toPromise();
+            expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                .toEqual(`
+        <igx-drop-down
+            (opening)="handleEvent($event, 'opening')"
+            (opened)="handleEvent($event, 'opened')"
+            (closing)="handleEvent($event, 'closing')"
+            (closed)="handleEvent($event, 'closed')"
+            (selecting)="handleEvent($event, 'selection')"
+        >
+            Display something onOpening, onClosing, onOpened, onClosed
+        </igx-drop-down>`);
+    });
+
+    // IgxToggleDirective
+    it('should update IgxToggleDirective event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update dropdown event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+            <div igxToggle
+                (onOpening)="eventHandler($event)"
+                (onAppended)="eventHandler($event)"
+                (onOpened)="eventHandler($event)"
+                (onClosing)="eventHandler($event)"
+                (onClosed)="eventHandler($event)"
+            >
+                <p>Some content that user would like to make it togglable.</p>
+            </div>`);
+                const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+                    .toPromise();
+
+                expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                    .toEqual(`
+            <div igxToggle
+                (opening)="eventHandler($event)"
+                (appended)="eventHandler($event)"
+                (opened)="eventHandler($event)"
+                (closing)="eventHandler($event)"
+                (closed)="eventHandler($event)"
+            >
+                <p>Some content that user would like to make it togglable.</p>
+            </div>`);
+    });
+
+    // IgxCombo
+    it('should update IgxCombo event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update combo event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+                <igx-combo
+                    (onSelectionChange)="eventHandler($event)"
+                    (onSearchInput)="eventHandler($event)"
+                    (onAddition)="eventHandler($event)"
+                    (onDataPreLoad)="eventHandler($event)"
+                    (onOpening)="eventHandler($event)"
+                    (onOpened)="eventHandler($event)"
+                    (onClosing)="eventHandler($event)"
+                    (onClosed)="eventHandler($event)"
+                >
+                </igx-combo>`);
+            const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+
+            expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                .toEqual(`
+                <igx-combo
+                    (selectionChange)="eventHandler($event)"
+                    (searchInputUpdate)="eventHandler($event)"
+                    (addition)="eventHandler($event)"
+                    (dataPreLoad)="eventHandler($event)"
+                    (opening)="eventHandler($event)"
+                    (opened)="eventHandler($event)"
+                    (closing)="eventHandler($event)"
+                    (closed)="eventHandler($event)"
+                >
+                </igx-combo>`);
+    });
+
+    // IgxSelect
+    it('should update IgxSelect event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update select event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+                <igx-select
+                    (onOpening)="eventHandler($event)"
+                    (onOpened)="eventHandler($event)"
+                    (onClosing)="eventHandler($event)"
+                    (onClosed)="eventHandler($event)"
+                >
+                </igx-select>`);
+            const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+
+            expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                .toEqual(`
+                <igx-select
+                    (opening)="eventHandler($event)"
+                    (opened)="eventHandler($event)"
+                    (closing)="eventHandler($event)"
+                    (closed)="eventHandler($event)"
+                >
+                </igx-select>`);
+    });
+
+    // IgxAutocomplete
+    it('should update IgxAutocomplete event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update autocomplete event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+            <input
+                igxInput
+                [igxAutocomplete]="townsPanel"
+                (onItemSelected)='itemSelected($event)'
+            />`);
+            const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+
+            expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                .toEqual(`
+            <input
+                igxInput
+                [igxAutocomplete]="townsPanel"
+                (itemSelected)='itemSelected($event)'
+            />`);
+    });
+
+    // IgxDialog
+    it('should update IgxDialog event subscriptions', () => {
+        pending('ts language service tests do not pass');
+    });
+
+    it('should update dialog event subscriptions in .html file', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html', `
+        <igx-dialog
+            (onOpen)="eventHandler($event)"
+            (onOpened)="eventHandler($event)"
+            (onClose)="eventHandler($event)"
+            (onClosed)="eventHandler($event)"
+            (onLeftButtonSelect)="eventHandler($event)"
+            (onRightButtonSelect)="eventHandler($event)"
+        >
+        </igx-dialog>`);
+            const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+
+            expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+                .toEqual(`
+        <igx-dialog
+            (opening)="eventHandler($event)"
+            (opened)="eventHandler($event)"
+            (closing)="eventHandler($event)"
+            (closed)="eventHandler($event)"
+            (leftButtonSelect)="eventHandler($event)"
+            (rightButtonSelect)="eventHandler($event)"
+        >
+        </igx-dialog>`);
+    });
+
 });
+
+