+++ projects/igniteui-angular/src/lib/drop-down/drop-down.common.ts
@@ -19,7 +19,7 @@ export const DropDownActionKey = mkenum({
 export type DropDownActionKey = (typeof DropDownActionKey)[keyof typeof DropDownActionKey];
 
 /**
- * Interface that encapsulates onSelection event arguments - old selection, new selection and cancel selection.
+ * Interface that encapsulates selecting event arguments - old selection, new selection and cancel selection.
  *
  * @export
  */
@@ -48,7 +48,7 @@ export const IGX_DROPDOWN_BASE = new InjectionToken<IDropDownBase>('IgxDropDownB
  * @hidden
  */
 export interface IDropDownList extends DisplayDensityBase {
-    onSelection: EventEmitter<ISelectionEventArgs>;
+    selecting: EventEmitter<ISelectionEventArgs>;
     width: string;
     height: string;
     id: string;
@@ -70,10 +70,10 @@ export interface IDropDownList extends DisplayDensityBase {
  */
 export interface IDropDownBase extends IDropDownList, IToggleView {
     selectedItem: any;
-    onOpening: EventEmitter<CancelableEventArgs>;
-    onOpened: EventEmitter<void>;
-    onClosing: EventEmitter<CancelableBrowserEventArgs>;
-    onClosed: EventEmitter<void>;
+    opening: EventEmitter<CancelableBrowserEventArgs>;
+    opened: EventEmitter<IBaseEventArgs>;
+    closing: EventEmitter<CancelableBrowserEventArgs>;
+    closed: EventEmitter<IBaseEventArgs>;
     allowItemsFocus?: boolean;
     setSelectedItem(index: number): void;
     selectItem(item: IgxDropDownItemBaseDirective, event?: Event): void;