+++ projects/igniteui-angular/src/lib/grids/grid-base.directive.ts
@@ -2659,11 +2659,6 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     /**
      * @hidden @internal
      */
-    public snackbarLabel = this.resourceStrings.igx_grid_snackbar_addrow_label;
-
-    /**
-     * @hidden @internal
-     */
     public calcWidth: number;
     /**
      * @hidden @internal