+++ projects/igniteui-angular/src/lib/select/select.component.ts
@@ -27,7 +27,7 @@ import { takeUntil } from 'rxjs/operators';
 import { DisplayDensityToken, IDisplayDensityOptions } from '../core/density';
 import { EditorProvider } from '../core/edit-provider';
 import { IgxSelectionAPIService } from '../core/selection';
-import { CancelableEventArgs, IBaseCancelableBrowserEventArgs, IBaseCancelableEventArgs, PlatformUtil } from '../core/utils';
+import { IBaseCancelableBrowserEventArgs, IBaseEventArgs, PlatformUtil } from '../core/utils';
 import { IgxLabelDirective } from '../directives/label/label.directive';
 import { IgxDropDownItemBaseDirective } from '../drop-down/drop-down-item.base';
 import { IGX_DROPDOWN_BASE, ISelectionEventArgs, Navigate } from '../drop-down/drop-down.common';
@@ -41,7 +41,7 @@ import { IgxSelectItemComponent } from './select-item.component';
 import { SelectPositioningStrategy } from './select-positioning-strategy';
 import { IgxSelectBase } from './select.common';
 import { IgxHintDirective, IgxInputGroupType, IGX_INPUT_GROUP_TYPE } from '../input-group/public_api';
-import { ToggleViewEventArgs } from '../directives/toggle/toggle.directive';
+import { ToggleViewCancelableEventArgs, ToggleViewEventArgs } from '../directives/toggle/toggle.directive';
 
 /** @hidden @internal */
 @Directive({
@@ -130,7 +130,7 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
      * ```
      */
     @Input()
-    overlaySettings: OverlaySettings;
+    public overlaySettings: OverlaySettings;
 
     /** @hidden @internal */
     @HostBinding('style.maxHeight')
@@ -140,41 +140,41 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
      * Emitted before the dropdown is opened
      *
      * ```html
-     * <igx-select onOpening='handleOpening($event)'></igx-select>
+     * <igx-select opening='handleOpening($event)'></igx-select>
      * ```
      */
     @Output()
-    public onOpening = new EventEmitter<IBaseCancelableEventArgs>();
+    public opening = new EventEmitter<IBaseCancelableBrowserEventArgs>();
 
     /**
      * Emitted after the dropdown is opened
      *
      * ```html
-     * <igx-select (onOpened)='handleOpened()'></igx-select>
+     * <igx-select (opened)='handleOpened($event)'></igx-select>
      * ```
      */
     @Output()
-    public onOpened = new EventEmitter<void>();
+    public opened = new EventEmitter<IBaseEventArgs>();
 
     /**
      * Emitted before the dropdown is closed
      *
      * ```html
-     * <igx-select (onClosing)='handleClosing($event)'></igx-select>
+     * <igx-select (closing)='handleClosing($event)'></igx-select>
      * ```
      */
     @Output()
-    public onClosing = new EventEmitter<IBaseCancelableBrowserEventArgs>();
+    public closing = new EventEmitter<IBaseCancelableBrowserEventArgs>();
 
     /**
      * Emitted after the dropdown is closed
      *
      * ```html
-     * <igx-select (onClosed)='handleClosed()'></igx-select>
+     * <igx-select (closed)='handleClosed($event)'></igx-select>
      * ```
      */
     @Output()
-    public onClosed = new EventEmitter<void>();
+    public closed = new EventEmitter<IBaseEventArgs>();
 
     /**
      * The custom template, if any, that should be used when rendering the select TOGGLE(open/close) button
@@ -376,7 +376,7 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
         }
 
         const args: ISelectionEventArgs = { oldSelection, newSelection, cancel: false };
-        this.onSelection.emit(args);
+        this.selecting.emit(args);
 
         if (args.cancel) {
             return;
@@ -426,7 +426,7 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
 }
 
     /** @hidden @internal */
-    ngAfterContentInit() {
+    public ngAfterContentInit() {
         this._overlayDefaults = {
             target: this.getEditElement(),
             modal: false,
@@ -450,11 +450,11 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
      *
      * @hidden @internal
      */
-    public handleOpening(event: CancelableEventArgs) {
-        const args: CancelableEventArgs = { cancel: event.cancel };
-        this.onOpening.emit(args);
+    public handleOpening(e: ToggleViewCancelableEventArgs) {
+        const args: IBaseCancelableBrowserEventArgs = { owner:this, event:e.event, cancel: e.cancel };
+        this.opening.emit(args);
 
-        event.cancel = args.cancel;
+        e.cancel = args.cancel;
         if (args.cancel) {
             return;
         }
@@ -472,20 +472,20 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
     /** @hidden @internal */
     public handleOpened() {
         this.updateItemFocus();
-        this.onOpened.emit();
+        this.opened.emit({ owner: this });
     }
 
     /** @hidden @internal */
-    public handleClosing(event) {
-        const args: CancelableEventArgs = { cancel: event.cancel };
-        this.onClosing.emit(args);
-        event.cancel = args.cancel;
+    public handleClosing(e: ToggleViewCancelableEventArgs) {
+        const args: IBaseCancelableBrowserEventArgs = { owner:this, event:e.event, cancel: e.cancel };
+        this.closing.emit(args);
+        e.cancel = args.cancel;
     }
 
     /** @hidden @internal */
     public handleClosed() {
         this.focusItem(false);
-        this.onClosed.emit();
+        this.closed.emit({ owner: this });
     }
 
     /** @hidden @internal */